PopulationAnalyzer.cpp
The "PopulationAnalyzer.cpp" program is a C++ application that allows users to manage and analyze population data of states and Union Territories (UTs) in India. It utilizes the C++ Standard Library's map container to store the population data of various states. Users can input the name of a state and its population, and then perform various operations on this data using a menu-driven interface.